CRUD for User, Patient, Role

+++ src/main/java/com/boots/entity/Role.java
@@ -9,7 +9,7 @@ import java.util.Set;
 @Table(name = "role")
 public class Role implements GrantedAuthority {
     @Id
-    private Long id;
+    private int id;
     private String name;
     @Transient
     @ManyToMany(mappedBy = "roles")
@@ -17,20 +17,20 @@ public class Role implements GrantedAuthority {
     public Role() {
     }
 
-    public Role(Long id) {
+    public Role(int id) {
         this.id = id;
     }
 
-    public Role(Long id, String name) {
+    public Role(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
